# src/services/swagger/swagger.yaml
openapi: 3.0.1
info:
  title: Swagger Catstore
  description: Friends and Cats
  termsOfService: http://swagger.io/terms/
  contact:
    email: alisphere.c@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: /v1
...
paths:
...
  /cat/{id}:
    get:
      tags:
        - cat
      summary: Find cat by ID
      description: Returns a single cat
      operationId: getCatById
      parameters:
        - name: id
          in: path
          description: ID of cat to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Cat'
            application/json:
              schema:
                $ref: '#/components/schemas/Cat'
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Cat not found
          content: {}
      security:
        - api_key: []
    post:
      tags:
        - cat
      summary: Updates a cat in the store with form data
      operationId: updateCatWithForm
      parameters:
        - name: id
          in: path
          description: ID of cat that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              properties:
                name:
                  type: string
                  description: Updated name of the cat
                status:
                  type: string
                  description: Updated status of the cat
      responses:
        405:
          description: Invalid input
          content: {}
      security:
        - catstore_auth:
            - write:cats
            - read:cats
    delete:
      tags:
        - cat
      summary: Deletes a cat
      operationId: deleteCat
      parameters:
        - name: api_key
          in: header
          schema:
            type: string
        - name: id
          in: path
          description: Cat id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Cat not found
          content: {}
      security:
        - catstore_auth:
            - write:cats
            - read:cats
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      xml:
        name: Category
...
    Tag:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      xml:
        name: Tag
...
    Cat:
      required:
        - name
        - photoUrls
      type: object
      properties:
        id:
          type: integer
          format: int64
        category:
          $ref: '#/components/schemas/Category'
        name:
          type: string
          example: doggie
        photoUrls:
          type: array
          xml:
            name: photoUrl
            wrapped: true
          items:
            type: string
        tags:
          type: array
          xml:
            name: tag
            wrapped: true
          items:
            $ref: '#/components/schemas/Tag'
        status:
          type: string
          description: cat status in the store
          enum:
            - available
            - pending
            - sold
      xml:
        name: Cat
...
  securitySchemes:
    catstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://dddd.mlml.lk
          scopes:
            write:cats: modify cats in your account
            read:cats: read your cats
    api_key:
      type: apiKey
      name: api_key
      in: header